requirements.txt
accelerate==1.10.0
antlr4-python3-runtime==4.9.3
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.1
fsspec==2024.6.1
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.9.79
omegaconf==2.3.0
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.0.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==70.2.0
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tqdm==4.67.1
transformers==4.55.3
triton==3.4.0
typing_extensions==4.12.2
urllib3==2.5.0
